--- game.go
@@ -1,8 +1,0 @@
-package main
-
-type Game struct {
-    Id      int     `json:"id"`
-    Name    string  `json:"name"`
-}
-
-type Games []Game
