+++ game/model.go
@@ -1,0 +1,8 @@
+package game
+
+type Game struct {
+	Id   int    `json:"id"`
+	Name string `json:"name"`
+}
+
+type Games []Game
