+++ main.go
@@ -1,20 +1,20 @@
-// Resources
-// https://thenewstack.io/make-a-restful-json-api-go/
-// https://semaphoreci.com/community/tutorials/building-and-testing-a-rest-api-in-go-with-gorilla-mux-and-postgresql
-
 package main
 
-import "os"
+//to add: curl -H "Content-Type: application/json" -d '{"name": "Camel Up"}' http://localhost:8080/games
+//to delete: curl -X "DELETE" http://localhost:8080/players/4
 
-var a App
+import (
+	"fmt"
+	"log"
+	"net/http"
+
+	_ "github.com/lib/pq"
+)
 
 func main() {
-    a = App{}
-    a.Initialize(
-        os.Getenv("GAPI_DB_USERNAME"),
-        os.Getenv("GAPI_DB_PASSWORD"),
-        os.Getenv("GAPI_DB_NAME"),
-    )
+	log.Fatal(http.ListenAndServe(":8080", NewRouter()))
+}
 
-    a.Run(":8080")
+func Index(w http.ResponseWriter, r *http.Request) {
+	fmt.Fprintf(w, "welcome to the gAPI")
 }
