--- player.go
@@ -1,8 +1,0 @@
-package main
-
-type Player struct {
-	Id		int		`json:"id"`
-	Name	string	`json:"name"`
-}
-
-type Players []Player
