+++ player/create.go
@@ -1,0 +1,48 @@
+package player
+
+import (
+	"encoding/json"
+	"io"
+	"io/ioutil"
+	"net/http"
+
+	"github.com/gapi/db"
+	"github.com/gapi/util"
+)
+
+func CreatePlayer(w http.ResponseWriter, r *http.Request) {
+	var p Player
+
+	data := db.GetDb()
+	body, err := ioutil.ReadAll(io.LimitReader(r.Body, 1048576))
+	if err != nil {
+		panic(err)
+	}
+	if errr := r.Body.Close(); errr != nil {
+		panic(errr)
+	}
+	if errrr := json.Unmarshal(body, &p); errrr != nil {
+		w.Header().Set("Content-Type", "application/json; charset=UTF-8")
+		w.WriteHeader(422) // unprocessable entity
+		if errrrr := json.NewEncoder(w).Encode(errrr); errrrr != nil {
+			panic(errrrr)
+		}
+	}
+	if errrrrr := p.CreatePlayer(data); errrrrr != nil {
+		util.RespondWithError(w, http.StatusInternalServerError, errrrrr.Error())
+		return
+	}
+	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
+	w.WriteHeader(http.StatusCreated)
+	if errrrrrr := json.NewEncoder(w).Encode(p); errrrrrr != nil {
+		util.RespondWithError(w, http.StatusInternalServerError, errrrrrr.Error())
+		return
+	}
+}
+
+// todo: figure out how to handle errors here
+func (p *Player) CreatePlayer(db *db.Database) error {
+	// fmt.Printf("getting here in createGame\n")
+	db.DB.Create(p)
+	return nil
+}
