+++ player/model.go
@@ -1,0 +1,11 @@
+package player
+
+import "github.com/jinzhu/gorm"
+
+type Players []Player
+
+type Player struct {
+	gorm.Model
+	Id   int
+	Name string
+}
